AS-2025-2026/5X-TPSIT/es-04_2026-02-04_superheroes-fastapi+sql-withoutORM/requirements.txt
httptools
uvicorn
fastapi
jinja2
mysql-connector-python